requirements.txt
streamlit
pandas
openpyxl
python-pptx
openai
python-dotenv
pytest
pytest-mock
lxml
